+++ RELEASE_NOTES.md
@@ -89,6 +89,7 @@ If you have problems upgrading to `7.0`, first try upgrading to `6.9`. It has a 
 * `KeyboardAction` `standardTextDocumentProxyAction` has been removed.
 * `KeyboardGesture` `.tap` has been removed, use `.release` instead.
 * `KeyboardInputViewController` replaces `actionCalloutContext` and `inputCalloutContext` with a single `calloutContext`.
+* `KeyboardInputViewController` `.shared` has been removed.
 * `KeyboardReturnActionMappable` has been refactored to `KeyboardActionMappable`.
 * `KeyboardTextField` now requires an keyboard input view controller.
 * `KeyboardTextView` now requires an keyboard input view controller.